BasicController/requirements.txt
# for odrivetool
odrive
pyusb
libusb
python-can[gs-usb]
matplotlib
pandas
numpy